requirements.txt
absl-py==2.3.1
accelerate==0.26.1
annotated-types==0.7.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
deepspeed==0.13.1
einops==0.6.1
filelock==3.20.0
fsspec==2025.10.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.45
grpcio==1.76.0
hf-xet==1.2.0
hjson==3.1.0
huggingface-hub==0.36.0
idna==3.11
joblib==1.5.2
kornia==0.6.12
Markdown==3.10
MarkupSafe==3.0.3
mpi4py==4.1.1
ninja==1.13.0
numpy==1.23.5
nvidia-ml-py==13.580.82
opencv-python==4.11.0.86
packaging==25.0
pandas==1.5.3
peft==0.10.0
pillow==12.0.0
platformdirs==4.5.0
protobuf==6.33.0
psutil==7.1.3
py-cpuinfo==9.0.0
pydantic==2.12.4
pydantic_core==2.41.5
pynvml==13.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.9.3
sentencepiece==0.1.99
sentry-sdk==2.43.0
shortuuid==1.0.13
six==1.17.0
smmap==5.0.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tokenizers==0.13.3
tqdm==4.67.1
transformers==4.31.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wandb==0.22.3
wcwidth==0.2.14
Werkzeug==3.1.3
